+++ packages/@webex/plugin-meetings/src/locus-info/selfUtils.ts
@@ -14,6 +14,7 @@ import {
   AUDIO,
   VIDEO,
   MediaContent,
+  SELF_ROLES,
 } from '../constants';
 import ParameterError from '../common/errors/parameter';
 
@@ -62,14 +63,16 @@ SelfUtils.parse = (self: any, deviceId: string) => {
       layout: SelfUtils.getLayout(self),
       canNotViewTheParticipantList: SelfUtils.canNotViewTheParticipantList(self),
       isSharingBlocked: SelfUtils.isSharingBlocked(self),
-      breakoutSessions: SelfUtils.getBreakouts(self),
+      breakoutSessions: SelfUtils.getBreakoutSessions(self),
+      breakout: SelfUtils.getBreakout(self),
     };
   }
 
   return null;
 };
 
-SelfUtils.getBreakouts = (self) => self?.controls?.breakout?.sessions;
+SelfUtils.getBreakoutSessions = (self) => self?.controls?.breakout?.sessions;
+SelfUtils.getBreakout = (self) => self?.controls?.breakout;
 
 SelfUtils.getLayout = (self) =>
   Array.isArray(self?.controls?.layouts) ? self.controls.layouts[0].type : undefined;
@@ -105,6 +108,7 @@ SelfUtils.getSelves = (oldSelf, newSelf, deviceId) => {
     current
   );
   updates.moderatorChanged = SelfUtils.moderatorChanged(previous, current);
+  updates.isUpgradeToModeratorOrCohost = SelfUtils.isUpgradeToModeratorOrCohost(previous, current);
   updates.isMediaInactiveOrReleased = SelfUtils.wasMediaInactiveOrReleased(previous, current);
   updates.isUserObserving = SelfUtils.isDeviceObserving(previous, current);
   updates.layoutChanged = SelfUtils.layoutChanged(previous, current);
@@ -141,13 +145,13 @@ SelfUtils.isJoined = (self: any) => self?.state === _JOINED_;
  *
  * @param {Self} previous - Previous self state
  * @param {Self} current - Current self state [per event]
- * @returns {boolean} - If the MEeting Layout Controls Layout has changed.
+ * @returns {boolean} - If the Meeting Layout Controls Layout has changed.
  */
 SelfUtils.layoutChanged = (previous: any, current: any) =>
   current?.layout && previous?.layout !== current?.layout;
 
 SelfUtils.breakoutsChanged = (previous, current) =>
-  !isEqual(previous?.breakoutSessions, current?.breakoutSessions);
+  !isEqual(previous?.breakoutSessions, current?.breakoutSessions) && !!current?.breakout;
 
 SelfUtils.isMediaInactive = (previous, current) => {
   if (
@@ -332,6 +336,31 @@ SelfUtils.moderatorChanged = (oldSelf, changedSelf) => {
   }
 
   return oldSelf.moderator !== changedSelf.moderator;
+};
+
+/**
+ * @param {Object} oldSelf
+ * @param {Object} changedSelf
+ * @returns {Boolean}
+ * @throws {Error} if changed self was undefined
+ */
+SelfUtils.isUpgradeToModeratorOrCohost = (oldSelf, changedSelf) => {
+  if (!oldSelf) {
+    return false;
+  }
+  if (!changedSelf) {
+    throw new ParameterError(
+      'New self must be defined to determine if self transitioned moderator or cohost status.'
+    );
+  }
+  const isAttendeeOnly =
+    oldSelf.roles.includes(SELF_ROLES.ATTENDEE) &&
+    !oldSelf.roles.includes(SELF_ROLES.COHOST) &&
+    !oldSelf.roles.includes(SELF_ROLES.MODERATOR);
+  const isCohost = changedSelf.roles.includes(SELF_ROLES.COHOST);
+  const isModerator = changedSelf.roles.includes(SELF_ROLES.MODERATOR);
+
+  return isAttendeeOnly && (isCohost || isModerator);
 };
 
 /**