+++ packages/@webex/plugin-meetings/test/unit/spec/meetings/utils.js
@@ -33,7 +33,7 @@ describe('plugin-meetings', () => {
         );
       });
 
-      it('should work fine if no default true site', () => {
+      it('should take the first site if none are default', () => {
         const userPreferences = {
           sites: [
             {
@@ -51,7 +51,10 @@ describe('plugin-meetings', () => {
           ],
         };
 
-        assert.equal(MeetingsUtil.parseDefaultSiteFromMeetingPreferences(userPreferences), '');
+        assert.equal(
+          MeetingsUtil.parseDefaultSiteFromMeetingPreferences(userPreferences),
+          'site1-example.webex.com'
+        );
       });
 
       it('should work fine if sites an empty array', () => {
@@ -60,6 +63,52 @@ describe('plugin-meetings', () => {
         };
 
         assert.equal(MeetingsUtil.parseDefaultSiteFromMeetingPreferences(userPreferences), '');
+      });
+    });
+
+    describe('#getThisDevice', () => {
+      it('return null if no devices in self', () => {
+        const newLocus = {};
+        assert.equal(MeetingsUtil.getThisDevice(newLocus), null);
+      });
+      it('return first device as this device', () => {
+        const newLocus = {
+          self: {
+            devices: [{state: 'JOINED'}]
+          }
+        };
+        assert.deepEqual(MeetingsUtil.getThisDevice(newLocus), {state: 'JOINED'});
+      })
+    });
+
+    describe('#joinedOnThisDevice', () => {
+      it('return false if no devices in self', () => {
+        const newLocus = {};
+        assert.equal(MeetingsUtil.joinedOnThisDevice(null, newLocus), false);
+      });
+      it('return true if joined on this device', () => {
+        const newLocus = {
+          self: {
+            devices: [{state: 'JOINED', correlationId: '111'}]
+          }
+        };
+        const meeting = {
+          correlationId: '111'
+        };
+
+        assert.equal(MeetingsUtil.joinedOnThisDevice(meeting, newLocus), true);
+      });
+      it('return true if selfMoved on this device', () => {
+        const newLocus = {
+          self: {
+            devices: [{state: 'LEFT', reason: 'MOVED', correlationId: '111'}]
+          }
+        };
+        const meeting = {
+          correlationId: '111'
+        };
+
+        assert.equal(MeetingsUtil.joinedOnThisDevice(meeting, newLocus), true);
       });
     });
   });